refactor: Add bytes2DataURL function for converting bytes to data URL
The code changes in fileUtils.py add a new function bytes2DataURL, which converts a given bytes object to a data URL. This function is useful for converting binary data, such as images, into a format that can be easily embedded in HTML or CSS. The function accepts an optional format parameter to specify the image format, with "PNG" being the default. This addition enhances the functionality of the fileUtils module and improves the flexibility of working with binary data.